requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
ffmpeg-python>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
PyJWT>=2.8.0
watchdog>=3.0.0

